Python-Backend/requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
